Update esotope.js
When https://abs.twimg.com/responsive-web/client-web/main.f6ffe885.js script is ran through Testcafe's Acorn and generated with this version of Esotope. An error occurs because `$props[0].value` is undefined, and `type` is trying to be retrieved from it as a object property. I've added in one small precaution to make sure this does not happen.